Migrate from react-scripts to vite
