--- src/views/components/Button/Button.js
@@ -1,22 +1,0 @@
-import React from "react";
-
-function Button(props) {
-    return (
-        <button><img src={props.icon} /></button>
-    )
-}
-
-Button.defaultProps = {
-    label: "",
-    fontStyle: "Calibri",
-    fontSize: "16px",
-    color: "#ffffff",
-    backgroundColor: null,
-    icon: null
-}
-
-Button.propTypes = {
-    label: PropTypes.string
-}
-
-export default Button;
